requirements.txt
# Scrapy 웹 크롤링 프레임워크
scrapy>=2.11.0

# Selenium 웹 브라우저 자동화
selenium>=4.15.0

# 웹드라이버 자동 관리
webdriver-manager>=4.0.0

# HTTP 요청 라이브러리
requests>=2.31.0

# HTML 파싱
beautifulsoup4>=4.12.0
lxml>=4.9.0

# 데이터 처리
pandas>=2.1.0
numpy>=1.24.0

# 날짜/시간 처리
python-dateutil>=2.8.0

# 로깅 및 유틸리티
colorama>=0.4.6

# === RAG 시스템 패키지 ===

# 임베딩 및 벡터 유사도
sentence-transformers>=2.2.2
transformers>=4.36.0
torch>=2.1.0

# 벡터 데이터베이스
chromadb>=0.4.18
faiss-cpu>=1.7.4

# LLM 및 AI 프레임워크
openai>=1.3.0
langchain>=0.0.340
langchain-openai>=0.0.2

# 텍스트 처리 (한국어)
konlpy>=0.6.0
kiwisolver>=1.4.0

# 데이터 시각화
matplotlib>=3.7.0
seaborn>=0.12.0

# 웹 프레임워크 (API 서버용)
fastapi>=0.104.0
uvicorn>=0.24.0

# 환경 변수 관리
python-dotenv>=1.0.0

# 유틸리티
tqdm>=4.66.0
pydantic>=2.5.0
psutil>=5.9.0 